# Conflicts:
#	src/main/java/noodlezip/common/config/SecurityConfig.java
#	src/main/java/noodlezip/community/controller/BoardController.java
#	src/main/java/noodlezip/community/service/BoardService.java
#	src/main/java/noodlezip/community/service/BoardServiceImpl.java
#	src/main/java/noodlezip/store/entity/Store.java

+++ src/main/java/noodlezip/store/dto/StoreDto.java
@@ -51,7 +51,7 @@ public class StoreDto implements Serializable {
     @NotNull
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private Integer storeLegalCode;
+    private Long storeLegalCode;
 
     // Store -> StoreDto
     public static StoreDto toDto(Store store) {
